Update Makefile and rbtree to include shared.c and fix types
Integrate shared.c into build targets and update mismatched types in rbtree functions for consistency with data sizes.

+++ rbtlib/rbtree.h
@@ -171,6 +171,6 @@ void listSharedMemoryEntities(const char *prefix);
 
 void createRbt(const int argc, char *argv[], void (*insertFunc)(Node **, FileInfo), const char *prefix);
 
-size_t getSharedMemorySize(const char *sharedMemoryName);
+long long getSharedMemorySize(const char *sharedMemoryName);
 
 #endif // RBTREE_H
